refactor test factories
