+++ src/tests/utils/factories.ts
@@ -1,123 +1,48 @@
-import { BigNumber } from '@ethersproject/bignumber'
-import { JsonRpcProvider } from '@ethersproject/providers'
-import { Wallet } from '@ethersproject/wallet'
-
+import { ChainId } from 'constants/chains'
+import { IndexSwapQuoteProvider } from 'quote'
 import {
-  FlashMintQuote,
-  FlashMintQuoteProvider,
-  FlashMintQuoteRequest,
-  SwapQuoteProvider,
-} from 'quote'
+  getLocalHostProviderUrl,
+  getZeroExSwapQuoteProvider,
+} from 'tests/utils'
+import { TestFactory } from 'tests/utils/test-factory'
 
-import { approveErc20, balanceOf } from './'
-
-class TxTestFactory {
-  constructor(readonly provider: JsonRpcProvider, readonly signer: Wallet) {}
-
-  /**
-   * Tests minting a given flash mint quote.
-   * @param quote a flash mint quote
-   */
-  async testMinting(quote: FlashMintQuote) {
-    const { signer } = this
-    const indexToken = quote.outputToken
-    const tx = quote.tx
-    if (!tx) fail()
-    const balanceBefore: BigNumber = await balanceOf(signer, indexToken.address)
-    if (quote.inputToken.symbol !== 'ETH') {
-      await approveErc20(
-        quote.inputToken.address,
-        quote.contract,
-        quote.inputOutputAmount,
-        signer
-      )
-    }
-    // Automatically adding from as it seems like estimateGas won't recognize
-    // the impersonated balance if `from` is not set.
-    tx.from = this.signer.address
-    const gasEstimate = await this.provider.estimateGas(tx)
-    tx.gasLimit = gasEstimate
-    const res = await signer.sendTransaction(tx)
-    res.wait()
-    const balanceAfter: BigNumber = await balanceOf(signer, indexToken.address)
-    expect(balanceAfter.gte(balanceBefore.add(quote.indexTokenAmount))).toBe(
-      true
-    )
-  }
-
-  /**
-   * Tests redeeming a given flash mint quote.
-   * @param quote a flash mint quote
-   * @param gasLimit override gas limit
-   */
-  async testRedeeming(quote: FlashMintQuote, gasLimit?: BigNumber) {
-    const { signer } = this
-    const indexToken = quote.inputToken
-    const balanceBefore: BigNumber = await balanceOf(signer, indexToken.address)
-    await approveErc20(
-      indexToken.address,
-      quote.contract,
-      quote.indexTokenAmount,
-      signer
-    )
-    const tx = quote.tx
-    if (!tx) fail()
-    // Automatically, adding from as it seems like estimateGas won't recognize
-    // the impersonated balance if `from` is not set.
-    tx.from = this.signer.address
-    tx.gasLimit = gasLimit
-    if (!gasLimit) {
-      const gasEstimate = await this.provider.estimateGas(tx)
-      tx.gasLimit = gasEstimate
-    }
-    const res = await signer.sendTransaction(tx)
-    res.wait()
-    const balanceAfter: BigNumber = await balanceOf(signer, indexToken.address)
-    expect(balanceAfter.lte(balanceBefore.sub(quote.indexTokenAmount))).toBe(
-      true
-    )
-  }
+// Pre-configured TestFactories
+export function getArbitrumTestFactory(
+  signer: any,
+  rpcUrl: string = getLocalHostProviderUrl(ChainId.Arbitrum)
+) {
+  const swapQuoteProvider = getZeroExSwapQuoteProvider(ChainId.Arbitrum)
+  return new TestFactory(rpcUrl, signer, swapQuoteProvider)
 }
 
-export class TestFactory {
-  private quote: FlashMintQuote | null = null
-  private quoteProvider: FlashMintQuoteProvider
-  private txFactory: TxTestFactory
-  constructor(
-    rpcUrl: string,
-    signer: Wallet,
-    swapQuoteProvider: SwapQuoteProvider
-  ) {
-    const provider = new JsonRpcProvider(rpcUrl)
-    this.quoteProvider = new FlashMintQuoteProvider(rpcUrl, swapQuoteProvider)
-    this.txFactory = new TxTestFactory(provider, signer)
-  }
+export function getArbitrumTestFactoryUniswap(
+  signer: any,
+  rpcUrl: string = getLocalHostProviderUrl(ChainId.Arbitrum)
+) {
+  const swapQuoteProvider = new IndexSwapQuoteProvider(rpcUrl)
+  return new TestFactory(rpcUrl, signer, swapQuoteProvider)
+}
 
-  async executeTx(gasLimit?: BigNumber) {
-    if (!this.quote) fail()
-    if (this.quote.isMinting) {
-      await this.txFactory.testMinting(this.quote)
-      return
-    }
-    await this.txFactory.testRedeeming(this.quote, gasLimit)
-  }
+export function getBaseTestFactory(
+  signer: any,
+  rpcUrl: string = getLocalHostProviderUrl(ChainId.Base)
+) {
+  const swapQuoteProvider = getZeroExSwapQuoteProvider(ChainId.Base)
+  return new TestFactory(rpcUrl, signer, swapQuoteProvider)
+}
 
-  async fetchQuote(config: FlashMintQuoteRequest): Promise<FlashMintQuote> {
-    const quote = await this.quoteProvider.getQuote(config)
-    expect(quote).toBeDefined()
-    if (!quote) fail()
-    expect(quote.isMinting).toEqual(config.isMinting)
-    expect(quote.indexTokenAmount).toEqual(config.indexTokenAmount)
-    expect(quote.inputOutputAmount.gt(0)).toBe(true)
-    this.quote = quote
-    return quote
-  }
+export function getMainnetTestFactory(
+  signer: any,
+  rpcUrl: string = getLocalHostProviderUrl(ChainId.Mainnet)
+) {
+  const swapQuoteProvider = getZeroExSwapQuoteProvider(ChainId.Mainnet)
+  return new TestFactory(rpcUrl, signer, swapQuoteProvider)
+}
 
-  getProvider(): JsonRpcProvider {
-    return this.txFactory.provider
-  }
-
-  getSigner(): Wallet {
-    return this.txFactory.signer
-  }
+export function getMainnetTestFactoryUniswap(
+  signer: any,
+  rpcUrl: string = getLocalHostProviderUrl(ChainId.Mainnet)
+) {
+  const swapQuoteProvider = new IndexSwapQuoteProvider(rpcUrl)
+  return new TestFactory(rpcUrl, signer, swapQuoteProvider)
 }
